archive_code/dqn_agent/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# dqn_agent/deadly_corridor/evaluate.py: 15
# dqn_agent/deadly_corridor/train.py: 17
# dqn_agent/dqn_common/agent.py: 10
# dqn_agent/dqn_common/game_wrapper.py: 10
# dqn_agent/dqn_common/replay_buffer.py: 11
# dqn_agent/dqn_common/utils.py: 9
# dqn_agent/simpler_basic/evaluate.py: 15
# dqn_agent/simpler_basic/train.py: 17
numpy == 1.19.5

# dqn_agent/dqn_common/utils.py: 8
opencv_python == 4.5.1.48

# dqn_agent/deadly_corridor/evaluate.py: 16
# dqn_agent/deadly_corridor/model.py: 14
# dqn_agent/deadly_corridor/train.py: 18
# dqn_agent/dqn_common/agent.py: 11
# dqn_agent/simpler_basic/evaluate.py: 16
# dqn_agent/simpler_basic/model.py: 14
# dqn_agent/simpler_basic/train.py: 18
tensorflow_gpu == 2.4.1
# or, just use tensorflow without gpu :)

# dqn_agent/dqn_common/game_wrapper.py: 9
vizdoom == 1.1.8
